pyproject.toml
[project]
name = "langgraph-agent-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Ezequiel Sirne",email = "ezequielsirne@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
dependencies = [
    "langchain (>=0.3.20,<0.4.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "openai (>=1.65.5,<2.0.0)",
    "streamlit (>=1.43.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
